+++ satellite/src/main/java/com/linbit/linstor/api/protobuf/ApplyStorPool.java
@@ -84,7 +84,6 @@ public class ApplyStorPool implements ApiCall
             null, // not needed by the satellite
             null,
             protoStorPool.getIsPmem(),
-            protoStorPool.getIsVdo(),
             protoStorPool.getIsExternalLocking()
         );
     }